+++ src/test/groovy/ru/vyarus/dropwizard/guice/debug/renderer/OptionsRendererTest.groovy
@@ -51,7 +51,7 @@ class OptionsRendererTest extends Specification {
         SearchCommands                 = false
         UseCoreInstallers              = true
         BindConfigurationByPath        = true
-        ConfigureFromGuiceModules      = true
+        AnalyzeGuiceModules            = true
         InjectorStage                  = PRODUCTION
         GuiceFilterRegistration        = [REQUEST]
         UseHkBridge                    = false
@@ -85,7 +85,7 @@ class OptionsRendererTest extends Specification {
         SearchCommands                 = false
         UseCoreInstallers              = true
         BindConfigurationByPath        = true
-        ConfigureFromGuiceModules      = true
+        AnalyzeGuiceModules            = true
         InjectorStage                  = PRODUCTION
         GuiceFilterRegistration        = [REQUEST]
         UseHkBridge                    = false
@@ -138,7 +138,7 @@ class OptionsRendererTest extends Specification {
         SearchCommands                 = false
         UseCoreInstallers              = true
         BindConfigurationByPath        = true
-        ConfigureFromGuiceModules      = true
+        AnalyzeGuiceModules            = true
         InjectorStage                  = PRODUCTION
         UseHkBridge                    = false